move pages and components folders into modules
