+++ packages/web/src/index.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import { render } from 'react-dom'
 
+// @ts-ignore
 import { RelayEnvironmentProvider } from '@entria/relay-experimental'
 import { Environment } from '@yotta/web/src/relay'
 
-import Main from './pages/Main'
+import Main from './modules/Main'
 
 render(
   <RelayEnvironmentProvider environment={Environment}>
