+++ packages/web/src/modules/Home/AskHeadline/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+// @ts-ignore
 import { useQuery } from '@entria/relay-experimental'
 import { graphql } from 'react-relay'
