+++ packages/web/src/modules/Navbar/YottaLogo/styles.ts
