+++ packages/web/src/routes.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { Router } from '@reach/router'
 
-import Home from './pages/Home'
-import Auth from './pages/Auth'
+import Home from './modules/Home'
+import Auth from './modules/Auth'
 
 const Routes = () => (
-  <Router className='router-container'>
-    <Home path='/' />
-    <Auth path='auth/github' />
+  <Router className="router-container">
+    <Home path="/" />
+    <Auth path="auth/github" />
   </Router>
 )
